python/mqtt_mysql_service/requirements.txt
mysqlclient==2.1.1
paho-mqtt==1.6.1
